+++ src/main/java/com/tajok/web/frontkit/service/IUserService.java
@@ -9,6 +9,6 @@ public interface IUserService {
 	public boolean loginCheck(String email ,String password);
 	public User getModel(String email);
 	public boolean registerCheck(String email, String password, String password_c);
-	public void register(String name, int phone, int level, String email, String password);
+	public void register(String name, long phone, int level, String email, String password);
 	public void delete(int id);
 }
